Changes added

+++ Backend-Marketplace/src/main/java/com/fcul/marketplace/dto/relatorio/RelatorioPorZonasDTO.java
@@ -9,11 +9,11 @@ public class RelatorioPorZonasDTO {
 
     //Key Representa o nome da zona
     //Value representa a quantidade de encomendas registadas na zona {key}
-    Map<String, Integer> mapEncomendasFreguesias;
-    Map<String, Integer> mapEncomendasMunicipio;
-    Map<String, Integer> mapEncomendasDistrito;
-    Map<String, Integer> mapEncomendasPais;
-    Map<String, Integer> mapEncomendasContinente;
+    Map<String, Long> mapEncomendasFreguesias;
+    Map<String, Long> mapEncomendasMunicipio;
+    Map<String, Long> mapEncomendasDistrito;
+    Map<String, Long> mapEncomendasPais;
+    Map<String, Long> mapEncomendasContinente;
     //Representa a quantidade de encomendas feitas no mundo
     Integer totalDeEncomendas;
 }
